pyproject.toml
[tool.poetry]
name = "weapon-app-backend"
version = "0.1.0"
description = ""
authors = ["Michael Vitollo <ace51689@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2.7"
requests = "^2.26.0"
django-cors-headers = "^3.9.0"
djangorestframework = "^3.12.4"
django-bootstrap4 = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
